Show running jobs above progress bar (#17)
Adds a small display showing running and failed jobs when running with
launch progress

Fixes a bug where crtl-c doesn't kill a workflow. This happens when a
tier launches more jobs than the scheduler has concurrency (which is
fine). The tier terminates and tries to kill its jobs via the websocket
connection, but skips over jobs which haven't haven't started yet and
thus don't have a WS connection. This results in the scheduler
continuing to launch new jobs and the tier left waiting for the
scheduler to finish them. I've added a hook to stop the scheduler from
launching new jobs, and modified the tier to set the complete event on
jobs it can't kill via websocket.

Fixes a bug in db_client resulting in the websocket connection never
being used.

Fixes a lint error for the dashboard code